no progress
tested a bunch of things to no avail:

mask_index = -1 doesn't work to disable collisions for some reason

tried both friction = 0 and physics_set_friction = 0, both of which had no effect on friction at all, which would've helped make the players fall off of the others' head

tried changing their collision mask hitboxes for idle, fwalk and bwalk to be elliptical so that they would naturally slide off due to physics interactions, but this frustratingly doesn't work either

going to need more time to figure this out, but committing the changes that I've done so far. likely going to revert everything but leaving this as a log of sorts.